Fix #2 raw_count is non-numeric when > 32k
If the table retrieved when calling get_table_link_lookup is
greater than 32,000 then the following error is produced:

  Error in raw_count/32000 : non-numeric argument to binary operator

raw_count is an object returned from httr::content and is
therefore non-numeric. Changing 'raw_count' to 'raw_count$count'
in the outer if statement resolves the error (and the test).

Note that there is another raw_count defined within the first clause
of the outer if statement which adds a little to the confusion and
perhaps ought to be renamed for clarity but the double use of
raw_count does not currently cause problems

All tests run and passed