Fix wrong offset number & incorrect null check in get_boundaries
The open geography portal allows batching of results and provides these
in 2000 record chunks. In the code this is incorrectly implemented as
200 so that there will be duplicate entries in the output. In addition,
the geography_filter null check was occuring after the geography_filter
had already been altered (line 30) and so was never triggered (even when
geography_filter was NULL). To solve this:

- The null check is moved to before the reassignment

- resultOffset is changed to 2000

- The logic is rearranged such that redundancy is removed and
  the url query logic is taken on by build_url_query_string

Tests all pass and the function is also much faster (one example
reduced from 48s to 3s)